requirements.txt
aldjemy==2.6
appdirs==1.4.4
apt-xapian-index==0.49
asgiref==3.4.1
backports.zoneinfo==0.2.1
blinker==1.4
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.1
django-js-asset==1.2.2
django-model-utils==4.2.0
django-simple-history==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
entrypoints==0.3
Flask==2.0.2
Flask-SQLAlchemy==2.5.1
gpg===1.13.1-unknown
greenlet==1.1.2
gunicorn==20.1.0
hibget==0.0.2
httplib2==0.14.0
idna==3.3
inflection==0.5.1
itsdangerous==2.0.1
itypes==1.2.0
Jinja2==3.0.3
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
MarkupSafe==2.0.1
mysqlclient==2.1.0
netifaces==0.10.4
oauth2-provider==0.0
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pbr==5.8.0
pexpect==4.6.0
Pillow==8.4.0
prompt-toolkit==3.0.23
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycairo==1.16.2
pycparser==2.21
pycrypto==2.6.1
pycups==1.9.73
PyGObject==3.36.0
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.6
PyQt5==5.14.1
python-apt==2.0.0+ubuntu0.20.4.6
python-debian===0.1.36ubuntu1
python-dotenv==0.19.2
python-pam==1.8.4
pytz==2021.3
PyYAML==5.3.1
reportlab==3.5.34
requests==2.27.1
requests-unixsocket==0.2.0
ruamel.yaml==0.17.20
ruamel.yaml.clib==0.2.6
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
sip==4.19.21
six==1.16.0
SQLAlchemy==1.4.31
sqlparse==0.4.2
systemd-python==234
testresources==2.0.1
ubuntu-advantage-tools==27.5
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.7
usb-creator==0.3.7
wadllib==1.3.3
wcwidth==0.2.5
Werkzeug==2.0.2
xkit==0.0.0
